+++ AWSS3/AWSS3Model.h
@@ -75,7 +75,9 @@ typedef NS_ENUM(NSInteger, AWSS3BucketLocationConstraint) {
     AWSS3BucketLocationConstraintUSGovEast1,
     AWSS3BucketLocationConstraintEUNorth1,
     AWSS3BucketLocationConstraintAPEast1,
-    AWSS3BucketLocationConstraintMESouth1
+    AWSS3BucketLocationConstraintMESouth1,
+    AWSS3BucketLocationConstraintAFSouth1,
+    AWSS3BucketLocationConstraintEUSouth1,
 };
 
 typedef NS_ENUM(NSInteger, AWSS3BucketLogsPermission) {