+++ CHANGELOG.md
@@ -2,16 +2,20 @@
 
 ## Unreleased
 
+### New features
+- **AWS Core**
+  - Added support for `af-south-1` - Africa (Cape Town) region
+  - Added support for `eu-south-1` - Europe (Milan) region
+
 ### Bug Fixes
 - **Amazon Pinpoint**
   - Fix Missing address issue related to apps which have enabled push notifications, is using the pinpoint SDK, but is not registering the token with the endpoint [PR: #2455](https://github.com/aws-amplify/aws-sdk-ios/pull/2455)
-
-## 2.13.2
 
 ### Misc. Updates
 - Model updates for the following services
   - Amazon EC2
   - Amazon Transcribe
+  - AWS Lambda
   
 ## 2.13.2